chore(deps): upgrade to postgres 18 and latest mattermost patch release
This revision bumps postgres from version 14 to the latest supported
release, version 18. In version 18, the format of the persistent storage
volume changed to improve ease of database upgrades. To accomodate this,
the volume mounts for the postgres service in docker compose was
adjusted. Refer to the upstream documentation for more details.

Additionally, mattermost was upgraded to the latest patch release
10.11.7 which includes some low- to medium-severity security fixes.